chore: separate error file
