+++ packages/shared/lib/core/profile-manager/errors/index.ts
@@ -2,4 +2,5 @@ export * from './cannot-remove-account.error'
 export * from './missing-transaction-progress-event-payload.error'
 export * from './remove-account-with-balance.error'
 export * from './remove-not-last-account.error'
-export * from './wallet-api-event.errors'
+export * from './wallet-api-event.error'
+export * from './wallet-api-event-validation.error'
