+++ packages/shared/lib/core/profile-manager/errors/wallet-api-event.error.ts
@@ -9,13 +9,3 @@ export class WalletApiEventError extends BaseError {
         })
     }
 }
-
-export class WalletApiEventValidationError extends BaseError {
-    constructor(reason: string) {
-        super({
-            message: reason,
-            ...DEFAULT_APP_ERROR_PARAMETERS,
-            localizeMessage: false,
-        })
-    }
-}
